main file init fixes - 12.17.20.1
changed some ".txt" to ".xml" since I am using xml and file extensions would change when they were saved. Also changed the face that a user could overwrite a main file by inputting the directory that the main file was in and not adding "main.xml" to the end of it. fixed this behavior, which required adding a new function to MusicFiles and some changes to Program